requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.
django-rest-swagger==2.2.0
drf-yasg==1.21.7
setuptools==69.5.1